+++ src/Collector/Stream/FilesystemStreamCollector.php
@@ -2,7 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Debug\Collector;
+namespace Yiisoft\Yii\Debug\Collector\Stream;
+
+use Yiisoft\Yii\Debug\Collector\CollectorInterface;
+use Yiisoft\Yii\Debug\Collector\CollectorTrait;
+use Yiisoft\Yii\Debug\Collector\IndexCollectorInterface;
 
 final class FilesystemStreamCollector implements CollectorInterface, IndexCollectorInterface
 {
